+++ src/main/java/ru/mail/polis/lsm/DAOFactory.java
@@ -1,6 +1,6 @@
 package ru.mail.polis.lsm;
 
-import ru.mail.polis.lsm.artem_drozdov.LsmDAO;
+import ru.mail.polis.lsm.eldar_tim.LsmDAO;
 
 import java.io.IOException;
